+++ receiver/src/statistics/statistics.cpp
@@ -42,7 +42,7 @@ uint64_t Statistics::GetTotalElapsedMs() const noexcept {
 
 
 void Statistics::SetWriteInterval(uint64_t interval_ms) {
-    write_interval_ = (size_t) interval_ms;
+    write_interval_ = static_cast<unsigned int>(interval_ms);
 }
 
 void Statistics::ResetStatistics() noexcept {
