+++ tests/automatic/producer/beamtime_metadata/beamtime_metadata.cpp
@@ -32,7 +32,7 @@ void ProcessCommandArguments(int argc, char* argv[], Args* args) {
     try {
         args->discovery_service_endpoint = argv[1];
         args->beamtime_id = argv[2];
-        args->mode = std::stoull(argv[3]);
+        args->mode = std::stoi(argv[3]);
         PrintCommandArguments(*args);
         return;
     } catch (std::exception& e) {
